+++ cli/src/commands/__tests__/install-test.js
@@ -35,7 +35,7 @@ import {
 const BASE_FIXTURE_ROOT = path.join(__dirname, '__install-fixtures__');
 
 function _mock(mockFn) {
-  return ((mockFn: any): JestMockFn<*, *>);
+  return ((mockFn: any): JestMockFn<any, any>);
 }
 
 async function touchFile(filePath) {
@@ -45,6 +45,14 @@ async function touchFile(filePath) {
 async function writePkgJson(filePath, pkgJson) {
   await fs.writeJson(filePath, pkgJson);
 }
+
+const defaultRunProps = {
+  overwrite: false,
+  verbose: false,
+  skip: false,
+  skipFlowRestart: true,
+  explicitLibDefs: [],
+};
 
 describe('install (command)', () => {
   describe('determineFlowVersion', () => {
@@ -123,6 +131,7 @@ describe('install (command)', () => {
           verbose: false,
           overwrite: false,
           skip: false,
+          skipCache: false,
           ignoreDeps: [],
           useCacheUntil: 1000 * 60,
         });
@@ -157,6 +166,7 @@ describe('install (command)', () => {
             verbose: false,
             overwrite: false,
             skip: false,
+            skipCache: false,
             ignoreDeps: [],
             useCacheUntil: 1000 * 60,
           });
@@ -187,6 +197,7 @@ describe('install (command)', () => {
           verbose: false,
           overwrite: false,
           skip: false,
+          skipCache: false,
           ignoreDeps: [],
           useCacheUntil: 1000 * 60,
         });
@@ -259,6 +270,7 @@ describe('install (command)', () => {
 
     const origConsoleLog = console.log;
     const origConsoleError = console.error;
+
     beforeEach(() => {
       (console: any).log = jest.fn();
       (console: any).error = jest.fn();
@@ -321,11 +333,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           ignoreDeps: [],
-          explicitLibDefs: [],
         });
 
         // Installs libdefs
@@ -400,11 +409,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           ignoreDeps: [],
-          explicitLibDefs: [],
         });
 
         // Installs libdefs
@@ -464,11 +470,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           ignoreDeps: ['dev', 'optional', 'bundled'],
-          explicitLibDefs: [],
         });
 
         // Installs libdefs
@@ -512,12 +515,7 @@ describe('install (command)', () => {
         ]);
 
         // Run the install command
-        await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
-          explicitLibDefs: [],
-        });
+        await run(defaultRunProps);
 
         // Installs a stub for someUntypedDep
         expect(
@@ -552,12 +550,7 @@ describe('install (command)', () => {
         ]);
 
         // Run the install command
-        await run({
-          overwrite: false,
-          verbose: false,
-          skip: true,
-          explicitLibDefs: [],
-        });
+        await run(defaultRunProps);
 
         // Installs a stub for someUntypedDep
         expect(
@@ -591,10 +584,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
+          ...defaultRunProps,
           overwrite: true,
-          verbose: false,
-          skip: false,
-          explicitLibDefs: [],
         });
 
         // Replaces the stub with the real typedef
@@ -630,12 +621,7 @@ describe('install (command)', () => {
         ]);
 
         // Run the install command
-        await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
-          explicitLibDefs: [],
-        });
+        await run(defaultRunProps);
 
         const libdefFilePath = path.join(
           FLOWPROJ_DIR,
@@ -650,12 +636,7 @@ describe('install (command)', () => {
         await fs.writeFile(libdefFilePath, libdefFileContent);
 
         // Run install command again
-        await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
-          explicitLibDefs: [],
-        });
+        await run(defaultRunProps);
 
         // Verify that the tweaked libdef file wasn't overwritten
         expect(await fs.readFile(libdefFilePath, 'utf8')).toBe(
@@ -683,12 +664,7 @@ describe('install (command)', () => {
         ]);
 
         // Run the install command
-        await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
-          explicitLibDefs: [],
-        });
+        await run(defaultRunProps);
 
         const libdefFilePath = path.join(
           FLOWPROJ_DIR,
@@ -703,10 +679,8 @@ describe('install (command)', () => {
 
         // Run install command again
         await run({
+          ...defaultRunProps,
           overwrite: true,
-          skip: false,
-          verbose: false,
-          explicitLibDefs: [],
         });
 
         // Verify that the tweaked libdef file wasn't overwritten
@@ -739,11 +713,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           packageDir: path.join(FLOWPROJ_DIR, '..'),
-          explicitLibDefs: [],
         });
 
         // Installs libdef
@@ -777,11 +748,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           rootDir: path.join(FLOWPROJ_DIR, 'src'),
-          explicitLibDefs: [],
         });
 
         // Installs libdef
@@ -829,11 +797,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           rootDir: path.join(FLOWPROJ_DIR, 'src'),
-          explicitLibDefs: [],
         });
 
         // Installs libdef
@@ -883,11 +848,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           rootDir: path.join(FLOWPROJ_DIR, 'src'),
-          explicitLibDefs: [],
         });
 
         // Installs libdef
@@ -949,11 +911,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           rootDir: path.join(FLOWPROJ_DIR, 'src'),
-          explicitLibDefs: [],
         });
 
         // Installs libdef
@@ -1003,11 +962,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           rootDir: path.join(FLOWPROJ_DIR, 'src'),
-          explicitLibDefs: [],
         });
 
         // Installs libdef
@@ -1046,11 +1002,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           rootDir: path.join(FLOWPROJ_DIR, 'src'),
-          explicitLibDefs: [],
         });
 
         // Installs libdef
@@ -1107,11 +1060,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           rootDir: path.join(FLOWPROJ_DIR, 'src'),
-          explicitLibDefs: [],
         });
 
         // Installs libdef
@@ -1169,11 +1119,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           rootDir: path.join(FLOWPROJ_DIR, 'src'),
-          explicitLibDefs: [],
         });
 
         // Installs libdef
@@ -1227,11 +1174,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           rootDir: path.join(FLOWPROJ_DIR, 'src'),
-          explicitLibDefs: [],
         });
 
         // Installs libdef
@@ -1257,6 +1201,48 @@ describe('install (command)', () => {
             ),
           ),
         ).toEqual(false);
+      });
+    });
+
+    it('does not crash with dependency listed as // for comments', () => {
+      return fakeProjectEnv(async FLOWPROJ_DIR => {
+        // Create some dependencies
+        await Promise.all([
+          mkdirp(path.join(FLOWPROJ_DIR, 'src')),
+          writePkgJson(path.join(FLOWPROJ_DIR, 'package.json'), {
+            name: 'test',
+            devDependencies: {
+              '//': 'some comment that should not install nor crash',
+              'flow-bin': '^0.43.0',
+            },
+            dependencies: {
+              foo: '1.2.3',
+            },
+          }),
+          mkdirp(path.join(FLOWPROJ_DIR, 'node_modules', 'flow-bin')),
+          mkdirp(path.join(FLOWPROJ_DIR, 'node_modules', 'foo')),
+        ]);
+
+        await touchFile(path.join(FLOWPROJ_DIR, 'src', '.flowconfig'));
+
+        // Run the install command
+        await run({
+          ...defaultRunProps,
+          rootDir: path.join(FLOWPROJ_DIR, 'src'),
+        });
+
+        // Installs libdef
+        expect(
+          await fs.exists(
+            path.join(
+              FLOWPROJ_DIR,
+              'src',
+              'flow-typed',
+              'npm',
+              'foo_v1.x.x.js',
+            ),
+          ),
+        ).toEqual(true);
       });
     });
   });
@@ -1322,11 +1308,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           ignoreDeps: [],
-          explicitLibDefs: [],
         });
 
         // Installs libdefs
@@ -1359,11 +1342,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           ignoreDeps: [],
-          explicitLibDefs: [],
         });
 
         // Installs libdefs
@@ -1393,11 +1373,8 @@ describe('install (command)', () => {
 
         // Run the install command
         await run({
-          overwrite: false,
-          verbose: false,
-          skip: false,
+          ...defaultRunProps,
           ignoreDeps: [],
-          explicitLibDefs: [],
         });
 
         // Installs libdefs