+++ cli/src/lib/npm/npmLibDefs.js
@@ -181,10 +181,15 @@ async function extractLibDefsFromNpmPkgDir(
   return libDefs;
 }
 
-async function getCacheNpmLibDefs(cacheExpiry) {
-  await ensureCacheRepo(cacheExpiry);
+export async function getCacheNpmLibDefs(
+  cacheExpiry: number,
+  skipCache: boolean = false,
+): Promise<Array<NpmLibDef>> {
+  if (!skipCache) {
+    await ensureCacheRepo(cacheExpiry);
+  }
   await verifyCLIVersion();
-  return getNpmLibDefs(path.join(getCacheRepoDir(), 'definitions'));
+  return await getNpmLibDefs(path.join(getCacheRepoDir(), 'definitions'));
 }
 
 const PKG_NAMEVER_RE = /^(.*)_v\^?([0-9]+)\.([0-9]+|x)\.([0-9]+|x)(-.*)?$/;
@@ -374,8 +379,11 @@ export async function findNpmLibDef(
   pkgVersion: string,
   flowVersion: FlowVersion,
   useCacheUntil: number = CACHE_REPO_EXPIRY,
+  skipCache?: boolean = false,
+  extLibDefs?: Array<NpmLibDef>,
 ): Promise<null | NpmLibDef> {
-  const libDefs = await getCacheNpmLibDefs(useCacheUntil);
+  const libDefs =
+    extLibDefs ?? (await getCacheNpmLibDefs(useCacheUntil, skipCache));
   const filteredLibDefs = filterLibDefs(libDefs, {
     type: 'exact',
     pkgName,