+++ definitions/npm/lodash_v4.x.x/flow_v0.55.x-v0.62.x/lodash_v4.x.x.js
@@ -2968,10 +2968,6 @@ declare module "lodash/fp" {
     ): any;
     unset(path: Path): (object: Object) => boolean;
     unset(path: Path, object: Object): boolean;
-    dissoc(path: Path): (object: Object) => boolean;
-    dissoc(path: Path, object: Object): boolean;
-    dissocPath(path: Path): (object: Object) => boolean;
-    dissocPath(path: Path, object: Object): boolean;
     update(
       path: Path
     ): ((updater: Function) => (object: Object) => Object) &